requirements.txt
pyotp==2.7.0
qrcode==7.3.1
Pillow==8.3.1
simplejson==3.17.6
pycryptodome==3.10.1